+++ frontend/src/pages/Tokenomics.tsx
@@ -1,0 +1,5 @@
+import { TokenomicsDashboard } from '../components/TokenomicsDashboard'
+
+export function Tokenomics() {
+  return <TokenomicsDashboard />
+}